Allow tools to handle arrow functions as class methods
Arrow functions are automatically bound to the class instance and can be
passed to child components. Normal functions need to be manually bound
before they are passed outside the instance, or the `this` variable will
be incorrect when called.

`babel-eslint` is required because eslint's default parser does not
implement experimental features.